+++ app/src/main/java/thanhnhan/myproject/socialmedia/ui/view/MainActivity.kt
@@ -223,167 +223,6 @@ fun MainApp(socketManager: SocketManager) {
                 })
             }
 
-            composable(
-                route = "choosePasswordSignUp/{email}",
-                arguments = listOf(
-                    navArgument("email") {
-                        type = NavType.StringType
-                    }
-                )
-            ) { backStackEntry ->
-                val email = backStackEntry.arguments?.getString("email")
-                requireNotNull(email)
-                ChoosePasswordSignUp(
-                    email = email,
-                    openChooseName = { email, password ->
-                        navController.navigate("chooseNameSignUp/$email/$password")
-                    },
-                    backAction = {
-                        navController.popBackStack()
-                    }
-                )
-            }
-
-            composable(
-                route = "chooseNameSignUp/{email}/{password}",
-                arguments = listOf(
-                    navArgument(name = "email") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "password") {
-                        type = NavType.StringType
-                    }
-                )
-            ) { navBackStackEntry ->
-                navBackStackEntry.arguments?.let { argument ->
-                    val email = argument.getString("email")!!
-                    val password = argument.getString("password")!!
-                    ChooseNameSignUp(
-                        email = email,
-                        password = password,
-                        openChooseBirthday = { email, password, name ->
-                            navController.navigate("chooseBirthdaySignUp/$email/$password/$name")
-                        },
-                        backAction = {
-                            navController.popBackStack()
-                        }
-                    )
-                }
-            }
-
-            composable(
-                route = "chooseBirthdaySignUp/{email}/{password}/{name}",
-                arguments = listOf(
-                    navArgument(name = "email") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "password") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "name") {
-                        type = NavType.StringType
-                    }
-                )
-            ) { navBackStackEntry ->
-                navBackStackEntry.arguments?.let { argument ->
-                    val email = argument.getString("email")!!
-                    val password = argument.getString("password")!!
-                    val name = argument.getString("name")!!
-                    ChooseBirthdaySignUp(
-                        email = email,
-                        password = password,
-                        name = name,
-                        openChooseCountry = { email, password, name, birthday ->
-                            navController.navigate("chooseCountrySignUp/$email/$password/$name/$birthday")
-                        },
-                        backAction = {
-                            navController.popBackStack()
-                        }
-                    )
-                }
-            }
-
-            composable(
-                route = "chooseCountrySignUp/{email}/{password}/{name}/{birthday}",
-                arguments = listOf(
-                    navArgument(name = "email") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "password") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "name") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "birthday") {
-                        type = NavType.StringType
-                    }
-                )
-            ) { navBackStackEntry ->
-                navBackStackEntry.arguments?.let { argument ->
-                    val email = argument.getString("email")!!
-                    val password = argument.getString("password")!!
-                    val name = argument.getString("name")!!
-                    val birthday = argument.getString("birthday")!!
-                    ChooseCountrySignUp(
-                        email = email,
-                        password = password,
-                        name = name,
-                        birthday = birthday,
-                        openVerifyCode = { email, password, name, birthday, country ->
-                            navController.navigate("verifyEmailCodeSignUp/$email/$password/$name/$birthday/$country")
-                        },
-                        openChooseEmail = {
-                            navController.navigate("chooseEmailSignUp")
-                        },
-                        backAction = {
-                            navController.popBackStack()
-                        }
-                    )
-                }
-            }
-
-            composable(
-                route = "verifyEmailCodeSignUp/{email}/{password}/{name}/{birthday}/{country}",
-                arguments = listOf(
-                    navArgument(name = "email") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "password") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "name") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "birthday") {
-                        type = NavType.StringType
-                    },
-                    navArgument(name = "country") {
-                        type = NavType.StringType
-                    }
-                )
-            ) { navBackStackEntry ->
-                navBackStackEntry.arguments?.let { argument ->
-                    val email = argument.getString("email")!!
-                    val password = argument.getString("password")!!
-                    val name = argument.getString("name")!!
-                    val birthday = argument.getString("birthday")!!
-                    val country = argument.getString("country")!!
-                    VerifyEmailCodeSignUp(
-                        email = email,
-                        password = password,
-                        name = name,
-                        birthday = birthday,
-                        country = country,
-                        openSignin = { email ->
-                            navController.navigate("signInScreen")
-                        },
-                        backAction = {
-                            navController.popBackStack()
-                        }
-                    )
-                }
-            }
             composable(route = "UserProfile") {
                 val context = LocalContext.current  // Lấy context bằng LocalContext
                 val dbHelper = UserDatabaseHelper(context)
